bean: Fix home page name in code
